Tests/LibWeb/Text/expected/abspos-relayout-selection-preserved.txt

partial-layouts=0 full-layouts=1
selection-painting-matches-full-layout=true
FAIL
